Add two ways to calculate `dirty`.
`OnChange` will immediately be dirty and stay dirty as soon as the value changes.

`NotEqual` will only be dirty if the value is differently than the defaulted `initValue`, and can change over time.

Add `defaultInitValue` because we're using this logic in 3 places now and I can see it changing in the future, since we may not want to default all null values to a string.